Bind to the data subfolder of the postgres container for FedLCM docker-compose deployment (#48)
The postgres image has a Volume item pointing to `/var/lib/postgresql/data`. If we don't explicitly bind this path, a volume will be automatically created by docker. And it will be destroyed after `docker-compose down`, causing data loss.

To address that, we explicitly use a host folder to bind to this path.

Signed-off-by: Fangchi Wang <wfangchi@vmware.com>